Refactoring of ocr codebase
