+++ ocr/model_48px_ctc.py
@@ -1,14 +1,9 @@
+import math
+from typing import List, Tuple, Optional
 
-from collections import defaultdict
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-
-import heapq
-import math
-import einops
-
-from typing import List, Tuple, Optional
 
 class PositionalEncoding(nn.Module):
 	def __init__(self, d_model, dropout=0.1, max_len=5000):
